export_requirements.txt
django>=5.0.2
psycopg2>=2.9.9
django-environ>=0.11.2
PyYAML>=6.0
